refactor: torna WithPagination um módulo reutilizável
