+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/ListContainer.tsx
@@ -1,3 +1,4 @@
+import { WithPagination } from "@/features/acf/common/frontend/WithPagination";
 import { WithSearch } from "@/features/acf/common/frontend/WithSearch";
 import React from "react";
 import type { AcfDashboardType } from "../../../../../../common/model";
@@ -8,7 +9,6 @@ import { FiltersBarCoeq } from "./modules/FiltersBar";
 import { InternalCardsCoeq } from "./modules/Internalcards";
 import { ToolBar } from "./modules/ToolBar";
 import { WithFilters } from "./modules/WithFilters";
-import { WithPagination } from "./modules/WithPagination";
 import { WithSorting } from "./modules/WithSorting";
 
 export type ListContainerProps = {
